requirements.txt
flask
playwright
google-api-python-client
google-auth
google-cloud-secret-manager